speed up computations on the approximative max-T null distribution
This reverts changes in r364 (0.4-14).  The savings obtained by
deferring the computation of the max-T null distribution, as in r364,
is practically negligible even in larger testing problems.  However,
the cost of computations _on_ the null distribution itself is
non-negligible.  As an example, consider the multcomp::adevent data.
Computing the null distribution for the 28-dimensional linear
statistic based on 5*10^6 resamples took 1.1x longer (31.9 vs 34.9
sec) when not deferring the max-T null distribution computation, but,
e.g., obtaining global and single-step adjusted p-values was shortened
by 3.8x (2.21 vs 0.58 sec) and 42.9x (30.9 vs 0.72 sec), respectively.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/coin@1773 edb9625f-4e0d-4859-8d74-9fd3b1da38cb